refactor(tg_client): split into modules
